aya-ebpf: use correct types for BPF helper return values
Complete the i32 errno migration started in 8962fc79. Errno-only
helpers now return Result<T, i32>, non-errno helpers (verdicts, byte
counts) return c_long, and internal functions receiving raw helper
returns use c_long.

Fixes: #1414

+++ ebpf/aya-ebpf/src/maps/bloom_filter.rs
@@ -16,18 +16,18 @@ impl<T> BloomFilter<T> {
     map_constructors!((), T, BPF_MAP_TYPE_BLOOM_FILTER, phantom _t);
 
     #[inline]
-    pub fn contains(&self, value: impl Borrow<T>) -> Result<(), i64> {
+    pub fn contains(&self, value: impl Borrow<T>) -> Result<(), i32> {
         let ret = unsafe {
             bpf_map_peek_elem(
                 self.def.as_ptr().cast(),
                 ptr::from_ref(value.borrow()).cast_mut().cast(),
             )
         };
-        (ret == 0).then_some(()).ok_or(ret)
+        (ret == 0).then_some(()).ok_or(ret as i32)
     }
 
     #[inline]
-    pub fn insert(&self, value: impl Borrow<T>, flags: u64) -> Result<(), i64> {
+    pub fn insert(&self, value: impl Borrow<T>, flags: u64) -> Result<(), i32> {
         let ret = unsafe {
             bpf_map_push_elem(
                 self.def.as_ptr().cast(),
@@ -35,6 +35,6 @@ impl<T> BloomFilter<T> {
                 flags,
             )
         };
-        (ret == 0).then_some(()).ok_or(ret)
+        (ret == 0).then_some(()).ok_or(ret as i32)
     }
 }
